+++ src/main/java/de/jeffclan/hooks/MinepacksHook.java
@@ -2,8 +2,10 @@ package de.jeffclan.hooks;
 
 import org.bukkit.Bukkit;
 import org.bukkit.inventory.Inventory;
+import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.Plugin;
 
+import at.pcgamingfreaks.Minepacks.Bukkit.API.Backpack;
 import at.pcgamingfreaks.Minepacks.Bukkit.API.MinepacksPlugin;
 import de.jeffclan.JeffChestSort.JeffChestSortPlugin;
 
@@ -17,7 +19,16 @@ public class MinepacksHook {
 	    Plugin bukkitPlugin = Bukkit.getPluginManager().getPlugin("Minepacks");
 	    if(plugin.hookMinepacks && bukkitPlugin instanceof MinepacksPlugin) {
 	        minepacks = (MinepacksPlugin) bukkitPlugin;
+			plugin.getLogger().info("Succesfully hooked into Minepacks");
 	    }
+	}
+	
+	public boolean isMinepacksBackpack(ItemStack item) {
+		if(minepacks == null) return false;
+		
+		if(minepacks.isBackpackItem(item)) return true;
+		
+		return false;
 	}
 	
 	public boolean isMinepacksBackpack(Inventory inv) {
@@ -25,22 +36,13 @@ public class MinepacksHook {
 		if(minepacks == null) return false;
 		
 		if( inv.getHolder() == null) return false;
-		
-		if( inv.getHolder().getClass().getName().equalsIgnoreCase("at.pcgamingfreaks.MinepacksStandalone.Bukkit.Backpack")) {
+				
+		if(inv.getHolder() instanceof Backpack) {
 			return true;
 		}
 		
 		return false;
 		
-		//System.out.println(inv.getHolder().getClass().getName());
-				
-		/*if(inv.getHolder() instanceof Backpack) {
-			return true;
-		}
-		
-		if(minepacks.getBackpackCachedOnly(p).getInventory() == inv) {
-			return true;
-		}*/
 	}
 
 }